Fix issues #39, #42, #43, #65, #67, #69
- #69: Object.values/entries returned NOT_FOUND markers (serialized as
  null) for number-keyed properties; integer ids must be looked up with
  the int-keyed get, not stringified first.
- #42: String.charCodeAt returned a Character, which behaves like a
  string in JS; it now returns a number.
- #43: Symbol.prototype.description implemented; Symbol() with no
  argument now has an undefined description instead of an empty string.
- #39: JSON.stringify no longer dumps internal class debug info for
  undefined, functions and symbols: they stringify to undefined at top
  level, are omitted from objects and become null in arrays (in both
  NativeJSON and NativeGSON). Plain JS objects are serialized via their
  property ids instead of the Map view, which could not distinguish
  undefined (must be skipped) from null (must be kept). Also fix number
  output: integral doubles printed a trailing .0 (invalid JS JSON
  output) and non-finite numbers now serialize as null.
- #67: NativeJavaMap threw ClassCastException when looking up JS
  property names on maps with non-String keys (e.g. TreeMap with
  comparable keys); property names are now converted to the map's key
  type and lookups that cannot succeed fall through to Java members.
- #65: interface adapters always invoked Java default methods; a JS
  object providing a function of the same name now overrides them.
  Also grant conversion weight for JS object -> non-functional
  interface so such interfaces are usable in method calls at all.

Includes regression tests for all of the above.

+++ src/main/java/dev/latvian/mods/rhino/NativeString.java
@@ -856,7 +856,8 @@ final class NativeString extends IdScriptableObject implements Wrapper {
 					if (id == Id_charAt) {
 						return String.valueOf(c);
 					}
-					return c;
+					// charCodeAt must return a Number, not a Character (which would behave like a string)
+					return (int) c;
 				}
 
 				case Id_indexOf: {
